addded models and managers

+++ sample_products_mobile_app/lib/utils/models/product.dart
@@ -11,7 +11,7 @@ class Product {
   late String? shortDescription;
   late String? description;
   late int? quantity;
-  late int? price;
+  late double? price;
   late int? brandId;
 
   late Brand? brand;
